Enhance doctor update functionality and validation
- Updated `DoctorsController` to handle doctor updates with new authorization and response types.
- Removed `Id` property from `UpdateDoctorResponse` to reflect new response structure.
- Changed return type of `UpdateDoctorAsync` in `DoctorService` and updated `IDoctorService` interface.
- Modified `AppUser` to use `int` for `Age` for broader value range.
- Cleaned up `QueryHandler` and enhanced `SpecificationEvaluator` for better querying.
- Added `UserFormValidator` for validating phone numbers and image extensions.
- Created `UpdateDoctorRequestValidator` for validating update requests.
- Generated initial database schema in migration files.
- Updated `ApplicationDbContextModelSnapshot` to reflect current model state.

+++ SmartClinic.Domain/Entities/AppUser.cs
@@ -18,7 +18,7 @@ public class AppUser : IdentityUser<int>
 
     public DateOnly BirthDate { get; set; }
 
-    public byte Age { get; private set; }
+    public int Age { get; private set; }
 
     public bool IsActive { get; set; } = true;
